refactor(saved-queries): extract savedQueryService from route handlers
Move all saved-query business logic (DB loaders, parameter resolution,
SQL safety pipeline, adapter lifecycle, execution) out of the route file
and into app/src/services/savedQueryService.js. Route handlers now parse
the request, call the service, and write the result envelope — matching
the AGENTS.md rule against business logic in routes.

Service uses the same {ok, statusCode, body} envelope as
queryOrchestrationService. No API contract changes; existing HTTP-level
tests pass unmodified.
